[Shubham]Added: validation and excepitons

+++ src/main/java/com/capgi/employeepayrollspring/service/EmployeePayrollService.java
@@ -2,6 +2,7 @@ package com.capgi.employeepayrollspring.service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +10,8 @@ import org.springframework.stereotype.Service;
 
 import com.capgi.employeepayrollspring.domain.EmployeePayrollDB;
 import com.capgi.employeepayrollspring.dto.EmployeePayrollDTO;
+import com.capgi.employeepayrollspring.exception.DataMissingException;
+import com.capgi.employeepayrollspring.exception.EmployeeNotFoundException;
 import com.capgi.employeepayrollspring.model.EmployeePayrollData;
 import com.capgi.employeepayrollspring.repository.EmployeePayrollRepository;
 
@@ -61,29 +64,66 @@ public class EmployeePayrollService implements IEmployeePayrollService {
 
 	@Override
 	public EmployeePayrollDTO getUserByIdDB(long id) {
-		return employeePayrollRepository.findById(id).map(employeePayroll -> new EmployeePayrollDTO(employeePayroll))
-				.orElse(null);
+		EmployeePayrollDTO employeePayrollDTO = employeePayrollRepository.findById(id)
+				.map(employeePayroll -> new EmployeePayrollDTO(employeePayroll)).orElse(null);
+		if (employeePayrollDTO.getName() != null) {
+			return employeePayrollDTO;
+		} else {
+			throw new EmployeeNotFoundException("Employee not found!!");
+		}
 	}
 
 	@Override
 	public EmployeePayrollDB createUserDB(EmployeePayrollDTO employeePayrollDTO) {
-		EmployeePayrollDB employeePayrollDB = new EmployeePayrollDB(employeePayrollDTO);
-		employeePayrollRepository.save(employeePayrollDB);
-		return employeePayrollDB;
+		if (employeePayrollDTO.getName() != null || employeePayrollDTO.getName() != "") {
+			EmployeePayrollDB employeePayrollDB = new EmployeePayrollDB(employeePayrollDTO);
+			employeePayrollRepository.save(employeePayrollDB);
+			return employeePayrollDB;
+		}
+		throw new EmployeeNotFoundException("Data is missing");
 	}
 
 	@Override
 	public EmployeePayrollDB updateUserDB(long empId, EmployeePayrollDTO empPayrollDTO) {
 		EmployeePayrollDB employeePayrollDB = employeePayrollRepository.findById(empId).get();
-		employeePayrollDB.setName(empPayrollDTO.getName());
-		employeePayrollDB.setSalary(empPayrollDTO.getSalary());
-		employeePayrollRepository.save(employeePayrollDB);
-		return employeePayrollDB;
+		if (empPayrollDTO.getName() == null) {
+			throw new DataMissingException("Data is not present!!");
+		} else {
+			if (employeePayrollDB != null && employeePayrollDB.getId() == empId) {
+				if (Objects.nonNull(empPayrollDTO.getName())) {
+					employeePayrollDB.setName(empPayrollDTO.getName());
+				}
+				if (Objects.nonNull(empPayrollDTO.getSalary())) {
+					employeePayrollDB.setSalary(empPayrollDTO.getSalary());
+				}
+				if (Objects.nonNull(empPayrollDTO.getDepartment())) {
+					employeePayrollDB.setDepartment(empPayrollDTO.getDepartment());
+				}
+				if (Objects.nonNull(empPayrollDTO.getGender())) {
+					employeePayrollDB.setGender(empPayrollDTO.getGender());
+				}
+				if (Objects.nonNull(empPayrollDTO.getStartDate())) {
+					employeePayrollDB.setStartDate(empPayrollDTO.getStartDate());
+				}
+				if (Objects.nonNull(empPayrollDTO.getNotes())) {
+					employeePayrollDB.setNotes(empPayrollDTO.getNotes());
+				}
+				employeePayrollRepository.save(employeePayrollDB);
+				return employeePayrollDB;
+			}
+		}
+		return null;
+
 	}
 
 	@Override
 	public void deleteUserDB(long empId) {
-		employeePayrollRepository.deleteById(empId);
+		EmployeePayrollDB employeePayrollDB = employeePayrollRepository.findById(empId).get();
+		if (employeePayrollDB.getName() != null) {
+			employeePayrollRepository.deleteById(empId);
+		} else {
+			throw new EmployeeNotFoundException("Employee not found!!!");
+		}
 	}
 
 }
